Add new genPoint buffer to avoid sampling margins
Previously we used the geopoint buffer to generate the vertices.
This caused us to generate vertices in the actual margin of the chunk
that was used for AO computations. The result of this was overlapping
meshes in the different chunks. This caused issues when we added correct
AO since the overlap created noise in. We solve this for now by having
another vertex buffer that is one smaller in XYZ than the geopoints
buffer